requirements.txt
groq==0.11.0
python-dotenv==1.0.1
openai==1.52.2
langchain-groq==0.2.0
beautifulsoup4==4.12.3
lxml==5.3.0